finally done

+++ data/posts/postsRouter.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const db = require('../../data/db');
-const { findById } = require('../../data/db');
+
 
 router.get('/', (req,res) => {
     try{
@@ -33,20 +33,16 @@ router.post('/', (req,res) => {
 
 })
 
-//-get ('/:id')-----//############################### error message to send no id
 router.get('/:id', (req,res) => {
     const {id} = req.params
     try{
         db.findById(id)
             .then( isthere => {
-                console.log(isthere)
-                if(isthere){
-                res.status(200).json(isthere) 
-                }else{
-                res.status(400).json({message:"no id found"})
+                console.log(isthere[0])
+                isthere[0] !== undefined?
+                    res.status(200).json(isthere):
+                    res.status(400).json({message:"no id found"})
                 }
-            }
-
             ).catch( err =>
                 res.status(400).json({message:"coulndt go through with it.."})
             )
@@ -55,18 +51,26 @@ router.get('/:id', (req,res) => {
     }
 
 })
-//--post ('/:id/comments')----//
+//--post ('/:id/comments')----// todo- id
 router.post('/:id/comments', (req,res) => {
     const {id} = req.params
     const comment = req.body
     comment.post_id = id
     try{
         if(comment.text){
-            db.insertComment(comment)
+            db.findById(id)
+                .then(foundid =>{
+                    if(foundid[0] !== undefined){
+                        db.insertComment(comment)
                         .then(com =>
                             res.status(201).json(comment))
                         .catch(err =>
                             res.status(400).json(err))
+                    }else{
+                        res.status(404).json({message:"id doesnt exist"})
+                    }
+                })
+                    
         }else{
             res.status(400).json({message:"add text field"})
         }
@@ -77,7 +81,7 @@ router.post('/:id/comments', (req,res) => {
     }
 
 })
-//--get ('/:id/comments')----//######################
+
 router.get('/:id/comments', (req,res) => {
     const {id} = req.params
     const comment = req.body
@@ -92,35 +96,42 @@ router.get('/:id/comments', (req,res) => {
         res.status(500).json({message:"The information could not be retrieved."});
     }
 })
-//--delete ('/:id')----//########################################
+
 router.delete('/:id', (req,res) => {
     const {id} = req.params
     try{
         db.findById(id)
-            .then( deletedobj =>
-                db.remove(id)
-                    .then(yesyes => console.log(deletedobj),
-                    res.status(200).json(deletedobj)
-                    ).catch( err=>
-                        res.status(200).json({message:"id doesnt exist"})
-                    )               
+            .then( deletedobj =>{
+                   if(deletedobj[0] !== undefined) {
+                        db.remove(id)
+                        .then(found =>
+                            res.status(200).json(deletedobj)
+                        ).catch( err=>
+                            res.status(200).json({message:"could not delete"}))
+                    }else{
+                        res.status(404).json({message:"id doesnt exist"})}}      
             ).catch( err =>
-                res.status(400).json({message:"coulndt go through with it.."})
-            )
-
+                res.status(400).json({message:"coulndt go through with it.."}))
     }catch{
         res.status(500).json({message:"The information could not be deleted."});
     }
-
 })
-//--put ('/:id')----//################################
+
 router.put('/:id', (req,res) => {
     const {id} = req.params
     const changes = req.body
 
     try{
-        db.update(id,changes)
-        .then(res.status(200).json(changes))
+        db.findById(id)
+        .then(foundid =>{
+            if(foundid[0] !== undefined){
+                db.update(id,changes)
+                .then(success => res.status(200).json(changes))
+                .catch(err => 
+                   res.status(400).json({message:"could not update"}) )
+            }else{
+                res.status(400).json({message:"id does not exist"})}
+        })
     }catch{
         res.status(500).json({message:`The informationat ${id}could not be edited.`});    
     }
